refactor(phase-base): narrow §19.3a scope + merge BootstrapContextBuilder
Address rubber-duck pass findings on Phase 12b validation:

* RED: §19.3a `Phases/Base/**/*.ts` glob was too broad — any future
  file added under the cluster would inherit the 30/20 cap without
  review. Narrowed to two file-scoped blocks covering EXACTLY the
  existing debt:
  - `BasePhase.ts` (422 LoC Template Method abstract class — keeps
    `max-lines: off` + 30/20 per-function).
  - `ActionContextBuilder.ts` (72 LoC file — only per-function
    exemption, `max-lines` drops back to the global cap). Phase 13
    drain target: collapse `buildActionContext` to ≤10 LoC and remove
    this entry.

* YELLOW: §7e (`import-x/prefer-default-export` off for `Phases/Base/**`)
  removed entirely. BootstrapContextBuilder.ts merged into its only
  consumer ActionContextBuilder.ts — `buildBootstrapContext` now lives
  alongside `buildActionContext` (the only call-site) so the pair share
  a single audit point with no separate file needed.

* YELLOW: Refactored `logHandoffSummary` (16 LoC) → extracted
  `resolveHandoffParts` helper using the project's Option type. Both
  functions now ≤10 LoC, matching the canonical cap.

* JSDoc cleanup: BalanceContextSlice.ts cross-reference updated to
  point at ActionContextBuilder.ts (sole owner after the merge).
  docs-coverage-allowlist.txt comment block updated from "five sibling
  files" to "four sibling files".

No public API change. `lib/index.cjs` exports unchanged
(`CompanyTypes`, `createScraper`, `SCRAPERS`). 72 targeted unit tests
pass: `npx jest --testPathPatterns="(Phases/Base|HandoffSummary|BasePhase|ActionContext|Timing|Architecture)"`.
