+++ src/Scrapers/Pipeline/Phases/Base/ActionContextBuilder.ts
@@ -3,14 +3,16 @@
  * access so a phase's ACTION stage can only reach the executor
  * surface and the read-only discovery slices. Browser-free phases
  * (INIT before launch, TERMINATE after teardown) fall back to
- * {@link "./BootstrapContextBuilder.ts" | buildBootstrapContext}.
+ * {@link buildBootstrapContext} (kept in this same module so the
+ * pair of builders share a single audit point — no §7e default-export
+ * exemption needed, no extra file).
  *
  * <p>Extracted from `Pipeline/Types/BasePhase.ts` during Phase 12b.
- * Uses an explicit object literal (NO spread) for the same audit
- * reason as its bootstrap sibling: silently inheriting future
- * IPipelineContext fields would defeat the seal-by-construction
- * contract that the TypeScript compiler relies on to reject
- * `resolveField` / `resolveVisible` from inside `action()`.
+ * Both builders use explicit object literals (NO spread) for audit
+ * reasons: silently inheriting future IPipelineContext fields would
+ * defeat the seal-by-construction contract that the TypeScript
+ * compiler relies on to reject `resolveField` / `resolveVisible`
+ * from inside `action()`.
  *
  * @see "../../Mediator/Elements/CreateElementMediator.ts" —
  *   {@link extractActionMediator} produces the executor surface.
@@ -18,9 +20,12 @@
 
 import { extractActionMediator } from '../../Mediator/Elements/CreateElementMediator.js';
 import { none, some } from '../../Types/Option.js';
-import type { IActionContext, IPipelineContext } from '../../Types/PipelineContext.js';
+import type {
+  IActionContext,
+  IBootstrapContext,
+  IPipelineContext,
+} from '../../Types/PipelineContext.js';
 import { balanceContextSlice } from './BalanceContextSlice.js';
-import { buildBootstrapContext } from './BootstrapContextBuilder.js';
 
 /**
  * Extract sealed executor from full context.
@@ -34,6 +39,41 @@ export function extractExecutor(ctx: IPipelineContext): IActionContext['executor
   const page = ctx.browser.value.page;
   const sealed = extractActionMediator(ctx.mediator.value, page);
   return some(sealed);
+}
+
+/**
+ * Build bootstrap context for INIT/TERMINATE — explicit object literal, NO spread.
+ * Has browser (for launch/teardown) but NO mediator, NO executor.
+ * Co-located with {@link buildActionContext} (the only consumer) so the
+ * pair share a single audit point.
+ * @param ctx - Full pipeline context.
+ * @returns IBootstrapContext with browser access.
+ */
+export function buildBootstrapContext(ctx: IPipelineContext): IBootstrapContext {
+  return {
+    options: ctx.options,
+    credentials: ctx.credentials,
+    companyId: ctx.companyId,
+    logger: ctx.logger,
+    diagnostics: ctx.diagnostics,
+    config: ctx.config,
+    fetchStrategy: ctx.fetchStrategy,
+    executor: none(),
+    apiMediator: ctx.apiMediator,
+    loginFieldDiscovery: ctx.loginFieldDiscovery,
+    preLoginDiscovery: ctx.preLoginDiscovery,
+    dashboard: ctx.dashboard,
+    scrapeDiscovery: ctx.scrapeDiscovery,
+    accountDiscovery: ctx.accountDiscovery,
+    txnEndpoint: ctx.txnEndpoint,
+    dashboardTxnHarvest: ctx.dashboardTxnHarvest,
+    authDiscovery: ctx.authDiscovery,
+    otpTrigger: ctx.otpTrigger,
+    api: ctx.api,
+    loginAreaReady: ctx.loginAreaReady,
+    ...balanceContextSlice(ctx),
+    browser: ctx.browser,
+  };
 }
 
 /**
